♻️ refactor: 컴포넌트 분리 및 form 검사 로직 추가
