+++ components/PostForm/PostForm.tsx
@@ -1,122 +1,74 @@
-import { useToast } from "@/contexts/ToastContext";
 import tooltipIcon from "@/public/images/tooltip-icon.svg";
-import { body1Regular, h1 } from "@/styles/fontsStyle";
+import { body1Regular } from "@/styles/fontsStyle";
 import styled from "@emotion/styled";
 import Image from "next/image";
-import { useRouter } from "next/navigation";
 import { ChangeEvent, useState } from "react";
-import Button from "../Button/Button";
 import FormContainer from "./FormContainer";
 import Tooltip from "./Tooltip";
 
-export default function PostForm({}) {
-  const router = useRouter();
-  const { showToast } = useToast();
+interface PostFormProps {
+  handleInputChange: (key: string, value: string) => void;
+}
 
+export default function PostForm({ handleInputChange }: PostFormProps) {
   const [isTooltipVisible, setIsTooltipVisible] = useState(false);
-  const [formData, setFormData] = useState({
-    wage: "",
-    start_at: "",
-    work_time: "",
-    description: "",
-  });
-
-  const handleInputChange = (key: string, value: string) => {
-    setFormData((prevData) => ({
-      ...prevData,
-      [key]: value,
-    }));
-  };
-
-  const handleRegisterClick = () => {
-    const confirmed = window.confirm(
-      `\n시급: ${formData.wage}\n시작 일시: ${formData.start_at}\n업무 시간: ${formData.work_time}\n공고 설명: ${formData.description}\n\n등록하시겠습니까?`,
-    );
-    if (confirmed) {
-      router.push("/"); // 공고 등록 한 후 id를 이용해 shops/${shopId}/notices/${noticeId}로 이동해야함
-      showToast("등록되었습니다!");
-    }
-  };
 
   return (
     <Wrapper>
-      <Title>공고 등록</Title>
-      <Content>
-        <TooltipIcon
-          src={tooltipIcon}
-          alt="tooltip_icon"
-          onMouseEnter={() => setIsTooltipVisible(true)}
-          onMouseLeave={() => setIsTooltipVisible(false)}
-        />
-        <Tooltip
-          visible={isTooltipVisible}
-          message="기존 시급과 같거나 상향된 금액을 입력하세요"
-        />
-        <FormContainer
-          label="시급*"
-          gridArea="wage"
-          inputProps={{
-            placeholder: "15,000",
-            unit: "원",
-            onChange: (event: ChangeEvent<HTMLInputElement>) =>
-              handleInputChange("wage", event.target.value),
-          }}
-        />
-        <FormContainer
-          label="시작 일시*"
-          gridArea="start_at"
-          inputProps={{
-            type: "date",
-            placeholder: "2023-07-01 15:00",
-            onChange: (event: ChangeEvent<HTMLInputElement>) =>
-              handleInputChange("start_at", event.target.value),
-          }}
-        />
-        <FormContainer
-          label="업무 시간*"
-          gridArea="work_time"
-          inputProps={{
-            placeholder: "3",
-            unit: "시간",
-            onChange: (event: ChangeEvent<HTMLInputElement>) =>
-              handleInputChange("work_time", event.target.value),
-          }}
-        />
-        <FormContainer
-          label="공고 설명"
-          gridArea="description"
-          inputProps={{
-            placeholder: "공고 내용",
-            onChange: (event: ChangeEvent<HTMLInputElement>) =>
-              handleInputChange("description", event.target.value),
-          }}
-        />
-      </Content>
-      <Button
-        text="등록하기"
-        color="colored"
-        width={312}
-        handleClick={handleRegisterClick}
+      <TooltipIcon
+        src={tooltipIcon}
+        alt="tooltip_icon"
+        onMouseEnter={() => setIsTooltipVisible(true)}
+        onMouseLeave={() => setIsTooltipVisible(false)}
+      />
+      <Tooltip
+        visible={isTooltipVisible}
+        message="기존 시급과 같거나 상향된 금액을 입력하세요"
+      />
+      <FormContainer
+        label="시급*"
+        gridArea="wage"
+        inputProps={{
+          placeholder: "15,000",
+          unit: "원",
+          onChange: (event: ChangeEvent<HTMLInputElement>) =>
+            handleInputChange("wage", event.target.value),
+        }}
+      />
+      <FormContainer
+        label="시작 일시*"
+        gridArea="start_at"
+        inputProps={{
+          type: "date",
+          placeholder: "2023-07-01 15:00",
+          onChange: (event: ChangeEvent<HTMLInputElement>) =>
+            handleInputChange("start_at", event.target.value),
+        }}
+      />
+      <FormContainer
+        label="업무 시간*"
+        gridArea="work_time"
+        inputProps={{
+          placeholder: "3",
+          unit: "시간",
+          onChange: (event: ChangeEvent<HTMLInputElement>) =>
+            handleInputChange("work_time", event.target.value),
+        }}
+      />
+      <FormContainer
+        label="공고 설명"
+        gridArea="description"
+        inputProps={{
+          placeholder: "공고 내용",
+          onChange: (event: ChangeEvent<HTMLInputElement>) =>
+            handleInputChange("description", event.target.value),
+        }}
       />
     </Wrapper>
   );
 }
 
 const Wrapper = styled.div`
-  background: var(--The-julge-gray-05, #fafafa);
-  display: flex;
-  padding: 60px;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
-  gap: 32px;
-`;
-
-const Title = styled.span`
-  ${h1}
-`;
-
-const Content = styled.div`
   position: relative;
   display: grid;
   width: 100%;
